requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Mail==0.10.0
Pillow==11.1.0
openpyxl==3.1.5
python-slugify==8.0.4
email-validator==2.2.0
requests==2.32.3
Flask-Limiter==3.8.0
gunicorn==23.0.0
